// ui/src/app/modules/administration/models/entitlement.ts
export class EntitlementViewModel {
    public claimCategoryData: ClaimDataModel[] = [];
    public roleAdminData: RoleCategoryDataModel[] = [];
    public categoryField = ['Activity Monitor', 'Admin/Workcenter', 'Plan & Schedule'];
    public selectedCategory: string;
}

export interface IClaimDataModel {
    id: number,
    code: string,
    description: string,
    overrideClaimId: number,
    category: string,
    isActive: boolean
}

export class ClaimDataModel implements IClaimDataModel {
    constructor(claimDataObj: IClaimDataModel) {
        this.id = claimDataObj.id;
        this.code = claimDataObj.code;
        this.description = claimDataObj.description;
        this.overrideClaimId = claimDataObj.overrideClaimId;
        this.category = claimDataObj.category;
        this.isActive = claimDataObj.isActive;
    }
    id: number;
    code: string;
    description: string;
    overrideClaimId: number;
    category: string;
    isActive: boolean;
}

export interface IRoleClaimMapList {
    id: number,
    code: string,
    description: string,
    overrideClaimId: number,
    category: string,
    isActive: boolean
}

export class RoleClaimMapList implements IRoleClaimMapList {
    constructor (roleClaimMapobj: IRoleClaimMapList) {
        this.id = roleClaimMapobj.id;
        this.code = roleClaimMapobj.code;
        this.description = roleClaimMapobj.description;
        this.overrideClaimId = roleClaimMapobj.overrideClaimId;
        this.category = roleClaimMapobj.category;
        this.isActive = roleClaimMapobj.isActive;
    }
    id: number;
    code: string;
    description: string;
    overrideClaimId: number;
    category: string;
    isActive: boolean;
}

export interface IRoleCategoryDataModel {
    id: number,
    name: string,
    isActive: boolean,
    description: string,
    mappedClaimsList: RoleClaimMapList[]
}

export class RoleCategoryDataModel implements IRoleCategoryDataModel {
    constructor (roleCategoryDataObj: IRoleCategoryDataModel) {
        this.id = roleCategoryDataObj.id;
        this.name = roleCategoryDataObj.name;
        this.isActive = roleCategoryDataObj.isActive;
        this.description = roleCategoryDataObj.description;
        this.mappedClaimsList = roleCategoryDataObj.mappedClaimsList;
    }
    id: number;
    name: string;
    isActive: boolean;
    description: string;
    mappedClaimsList: RoleClaimMapList[] = [];
}

export interface IRoleCategoryReq {
    id: number,
    name: string,
    description: string,
    mappedClaims: number[],
    mappedCMMSRoles: number[],
    isActive: boolean,
    createdBy: string,
    createdDateTime: string,
    lastUpdatedBy: string,
    lastUpdatedDateTime: string

}

export class RoleCategoryReq implements IRoleCategoryReq {
    
    constructor (roleCategoryReqObj: IRoleCategoryReq) {
        this.id = roleCategoryReqObj.id;
        this.name = roleCategoryReqObj.name;
        this.description = roleCategoryReqObj.description;
        this.mappedClaims = roleCategoryReqObj.mappedClaims;
        this.mappedCMMSRoles = roleCategoryReqObj.mappedCMMSRoles;
        this.isActive = roleCategoryReqObj.isActive;
        this.createdBy = roleCategoryReqObj.createdBy;
        this.createdDateTime = roleCategoryReqObj.createdDateTime;
        this.lastUpdatedBy = roleCategoryReqObj.lastUpdatedBy;
        this.lastUpdatedDateTime = roleCategoryReqObj.lastUpdatedDateTime;
    }
    id: number;
    name: string;
    description: string;
    mappedClaims: number[];
    mappedCMMSRoles: number[];
    isActive: boolean;
    createdBy: string;
    createdDateTime: string;
    lastUpdatedBy: string;
    lastUpdatedDateTime: string;
}